requirements-canslim.txt
python-dotenv==1.0.1
yfinance==0.2.38
pandas==2.2.2
numpy==1.26.4
slack-sdk==3.27.1
schedule==1.2.1
ta==0.11.0 